Split executor/aggregate.rs into two modules,
Add struct State which manages state during aggregating.
Let State to manage "index" which was not proper to locate in main aggregate function.
